fix(tools): rebuild hero order with observed removals
