+++ apps/tools/src/host.test.ts
@@ -21,7 +21,6 @@ describe("native Tools host diagnostics", () => {
       setPlayerSkills: command,
       setPlayerAttributes: command,
       addHero: command,
-      kickAllHeroes: command,
       kickHero: command,
       setHeroBehaviour: command,
       setHeroSecondary: command,
@@ -70,7 +69,6 @@ describe("native Tools host diagnostics", () => {
       setPlayerSkills: command,
       setPlayerAttributes: command,
       addHero: command,
-      kickAllHeroes: command,
       kickHero: command,
       setHeroBehaviour: command,
       setHeroSecondary: command,
@@ -124,7 +122,6 @@ describe("native Tools host diagnostics", () => {
       setPlayerSkills: command,
       setPlayerAttributes: command,
       addHero: command,
-      kickAllHeroes: command,
       kickHero: command,
       setHeroBehaviour: command,
       setHeroSecondary: command,
